third_party/blink/web_tests/http/tests/loading/fire-error-event-empty-404-script-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
Test for bug 50589: Fire an error event for empty 404 script.

Should say SUCCESS:

SUCCESS
